Set up template admin to use edit slots inline
For now, the inline solution is not only more Django-like but much
easier than figuring out how to completely replace the admin view and
posting some custom format for all the data.

In the future, I'd like to build a custom inline template for a nice,
pretty calendar, but that's very low priority.